useReducer explained
useReducer is a React hook that allows you to manage complex states in functional components in a scalable way. It uses the concept of reduction, where a reducer function is used to update the state.

In the provided code, useReducer is used to manage the state of a form field. The initial state is defined as an object that contains a value and an isTouched property that indicates whether the field has been touched or not. The reducer function inputStateReducer is defined to update the state of the form field based on different types of actions.

The useInput function uses useReducer to manage the state of the form field. It receives a validateValue function as an argument, which is used to validate the value of the form field.

The useInput function returns an object with different properties and methods that can be used by the component. These properties include the current value of the form field, whether it is valid or not, whether there is an error, and the event handlers for when the value of the field changes or the field is touched. The reset method is used to reset the state of the form field.

In summary, useReducer is a very useful hook for managing complex states in functional components. It allows you to define a reducer function that updates the state of the component based on different actions, making the code more scalable and easy to manage.